prevent positions table if not party id provided, handle no data in async renderer
