+++ libs/positions/src/lib/positions-container.tsx
@@ -1,69 +1,17 @@
-import { useRef, useCallback, useMemo } from 'react';
-import { produce } from 'immer';
-import merge from 'lodash/merge';
-import { useRouter } from 'next/router';
-import { AsyncRenderer } from '@vegaprotocol/ui-toolkit';
-import { useDataProvider } from '@vegaprotocol/react-helpers';
+import { Splash } from '@vegaprotocol/ui-toolkit';
 import { useVegaWallet } from '@vegaprotocol/wallet';
-import { PositionSubscribe_positions } from './__generated__/PositionSubscribe';
-import { Positions_party_positions } from './__generated__/Positions';
-
-import type { AgGridReact } from 'ag-grid-react';
-import PositionsTable, { getRowNodeId } from './positions-table';
-import { positionsDataProvider } from './positions-data-provider';
+import { PositionsManager } from './positions-manager';
 
 export const PositionsContainer = () => {
-  const { pathname, push } = useRouter();
-  const gridRef = useRef<AgGridReact | null>(null);
   const { keypair } = useVegaWallet();
-  const variables = useMemo(() => ({ partyId: keypair?.pub }), [keypair]);
-  const update = useCallback(
-    (delta: PositionSubscribe_positions) => {
-      const update: Positions_party_positions[] = [];
-      const add: Positions_party_positions[] = [];
-      if (!gridRef.current) {
-        return false;
-      }
-      const rowNode = gridRef.current.api.getRowNode(getRowNodeId(delta));
-      if (rowNode) {
-        const updatedData = produce<Positions_party_positions>(
-          rowNode.data,
-          (draft: Positions_party_positions) => {
-            merge(draft, delta);
-          }
-        );
-        if (updatedData !== rowNode.data) {
-          update.push(updatedData);
-        }
-      } else {
-        add.push(delta);
-      }
-      if (update.length || add.length) {
-        gridRef.current.api.applyTransactionAsync({
-          update,
-          add,
-          addIndex: 0,
-        });
-      }
-      return true;
-    },
-    [gridRef]
-  );
-  const { data, error, loading } = useDataProvider<
-    Positions_party_positions,
-    PositionSubscribe_positions
-  >(positionsDataProvider, update, variables);
-  return (
-    <AsyncRenderer loading={loading} error={error} data={data}>
-      {(data) => (
-        <PositionsTable
-          ref={gridRef}
-          data={data}
-          onRowClicked={(id) =>
-            push(`${pathname}/${id}?portfolio=orders&trade=orderbook`)
-          }
-        />
-      )}
-    </AsyncRenderer>
-  );
+
+  if (!keypair) {
+    return (
+      <Splash>
+        <p>Please connect Vega wallet</p>
+      </Splash>
+    );
+  }
+
+  return <PositionsManager partyId={keypair.pub} />;
 };
