+++ libs/positions/src/lib/positions-table.tsx
@@ -15,7 +15,6 @@ import { MarketTradingMode } from '@vegaprotocol/types';
 
 interface PositionsTableProps {
   data: Positions_party_positions[] | null;
-  onRowClicked: (marketId: string) => void;
 }
 
 export const getRowNodeId = (data: { market: { id: string } }) =>
@@ -47,7 +46,7 @@ interface PositionsTableValueFormatterParams extends ValueFormatterParams {
 }
 
 export const PositionsTable = forwardRef<AgGridReact, PositionsTableProps>(
-  ({ data, onRowClicked }, ref) => {
+  ({ data }, ref) => {
     const sortedData = useMemo(() => {
       return compact(data).sort(sortByName);
     }, [data]);
@@ -62,9 +61,6 @@ export const PositionsTable = forwardRef<AgGridReact, PositionsTableProps>(
           flex: 1,
           resizable: true,
         }}
-        onRowClicked={({ data }: { data: Positions_party_positions }) =>
-          onRowClicked(getRowNodeId(data))
-        }
         components={{ PriceCell }}
       >
         <AgGridColumn
